Add an "output pipeline" stage, that is where final int32 accumulators are converted into the final uint8 values and stored into the destination matrix.
The existing API and ABI are preserved, the Gemm function is untouched.
There is a new public entry point, GemmWithOutputPipeline, allowing to
use a custom output pipeline instead of the default one (which
quantizes values down to 8 bit, clamps them, and casts to uint8).

A pipeline is just a std::tuple of "stages", which are structs defined
in public/output_stages.h.

The intended purposes are:
  1. Allow getting the raw int32 result. (Just pass an empty tuple).
  2. Allow fusing other output operations, such as bias additions,
     activation functions, etc. (One will pass longer tuples with
     such stages.). Those will come in a subsequent commit.