Generalize usage of correct rounding shifts in output stages.
This only makes a difference for negative values.

Consequently, the legacy output pipeline (as exposed by Gemm, EightBitIntGemm
entry points) is unaffected. Indeed, since it adds the result_offset
before shifting, and is immediately followed by clamping to [0,255],
the behavior of the rounding shift on negative values is
irrelevant to it. That is why the ReferenceEightBitIntGemm function
in test.cc did not need to be updated.

On the other hand, more general output pipelines are affected. In principle,
it is the task of TestOutputStages to test them. Unfortunately, that function
is not in such a good shape at the moment; for instance, the hard-coded
offsets
  const int lhs_offset = 12;
  const int rhs_offset = -34;
at test.cc:1188 are not large-negative enough to actually produce negative
accumulators and thus exercise the nuance being addressed here. I took a
look at fixing this function and fixed what could be fixed without a
complete rewrite, but ultimately this needs a rewrite to take a fully
principled approach to generating quantized GEMM call patterns from
the quantization of a float GEMM workload, as illustrated in
doc/quantization_example.cc.