Remove the old requantizing less-than-8-bit stuff.
Map the DefaultL7R5BitDepthParams setting to DefaultL8R8BitDepthParams
so that this change does not break any user relying on it, since
the whole point of requantization was to make this an implementation
detail.

Instead, from now on, using actual less-than-8-bit settings will only
have the effect of selecting a different GEMM kernel. In other words,
it is now the responsibility of the user to know the actual bit-depth
(i.e. range) of their 8bit values. In exchange for that responsibility,
users of lower-than-8-bit depths will enjoy the increased performance
of corresponding kernels without the overhead of requantization.