Do not round up the accumulation loop size more than necessary (round up to next multiple of kernel depth, not to L2 block depth).
This exposed a bug in SSE kernels, which were using a JNZ instruction
to close the peeled-by-2 loop, thereby assuming the loop size to be
a multiple of twice the kernel depth. It is only correct to assume
it to be a multiple of the kernel depth. Fixed by using JA instead
of JNZ.