New fast kernel on ARM64, associated with new BitDepthParam: L8R8WithLhsNonzeroBitDepthParams
The way that this kernel is faster, is that it treats its 8bit operands
as signed int8 and takes advantage of the fact that a product of int8
values is in the [-2^14 ; 2^14] range, therefore it is *almost* safe
to add two of them in a int16 accumulator.

Unfortunately that is not always safe: -128*-128 == 2^14, and adding
that value twice would overflow int16.

Therefore, we cannot just switch everyone's default kernel to that.
Instead, we add that new BitDepthParam whereby the user may specify
that one of the operands will never take the minimum representable 8bit
value. Then gemmlowp can pick up this kernel.

Note that I/O 8-bit values remain uint8, so when using a int8 kernel,
there is some work in the pack/unpack stages to bridge that difference.
The pack stage has to XOR the top bit to convert uint8 input values
to int8 values (e.g. 0 becomes -128, 255 becomes 127). The unpack stage
has to correspondingly add 128 to the lhs_offset, rhs_offset.

The test is cleaned up and refactored a bit to consistently test
all BitDepthParams, etc.